has finished adding and started editing

+++ shop-service/src/main/java/springboot/dao/RequestRepository.java
@@ -3,6 +3,6 @@ package springboot.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import springboot.model.Request;
 
-public interface RequestRepository extends JpaRepository<Request, Integer> {
+public interface RequestRepository extends JpaRepository<Request, Long> {
 
 }
